TS-migration: ListItem
