+++ src/components/ListItem/__stories__/ListItem.stories.mdx
@@ -1,5 +1,5 @@
 import ListItem from "../ListItem";
-import { ArgsTable, Story, Canvas, Meta } from "@storybook/addon-docs";
+import { ArgsTable, Canvas, Meta, Story } from "@storybook/addon-docs";
 import { createComponentTemplate, createStoryMetaSettings } from "../../../storybook/functions/create-component-story";
 import Send from "../../Icon/Icons/components/Send";
 import ListItemIcon from "../../ListItemIcon/ListItemIcon";
@@ -40,7 +40,7 @@ An item of a List component
 </Canvas>
 
 ## Props
-<ArgsTable of={ ListItem } />
+<ArgsTable story="Overview" />
 
 ## Usage
 <UsageGuidelines guidelines={[
